refactor: Dealer와 PlayerGroup을 관리하기 위한 GamerGroup 생성
