+++ src/main/java/blackjack/domain/BlackJack.java
@@ -1,29 +1,22 @@
 package blackjack.domain;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.EnumMap;
 import java.util.List;
-import java.util.Map;
 
 public class BlackJack {
-    private final Dealer dealer;
-    private final PlayerGroup playerGroup;
+    private final GamerGroup gamerGroup;
     private final CardPack cardPack;
 
     public BlackJack(PlayerGroup playerGroup) {
-        this.playerGroup = playerGroup;
-        this.dealer = new Dealer();
+        this.gamerGroup = new GamerGroup(new Dealer(), playerGroup);
         this.cardPack = new CardPack();
     }
 
     public void divideCards() {
-        playerGroup.addTwoCards(cardPack);
-        dealer.addTwoCards(cardPack.pickOne(), cardPack.pickOne());
+        gamerGroup.addTwoCards(cardPack);
     }
 
     public Dealer getDealer() {
-        return dealer;
+        return gamerGroup.getDealer();
     }
 
     public void addCardTo(Player player) {
@@ -31,46 +24,18 @@ public class BlackJack {
     }
 
     public GameResult getGameResult() {
-        Map<String, Match> playerResults = playerGroup.getPlayerResult(dealer.getMaxCardGroupSum());
-        PlayerResult playerResult = new PlayerResult(playerResults);
-
-        Collection<Match> playerMatches = playerResults.values();
-        Map<Match, Integer> dealerMatches = initializeDealerResults(playerMatches);
-        DealerResult dealerResult = new DealerResult(dealer.getName(), dealerMatches);
-
-        return new GameResult(dealerResult, playerResult);
+        return gamerGroup.getGameResult();
     }
 
-    private Map<Match, Integer> initializeDealerResults(Collection<Match> matches) {
-        Map<Match, Integer> matchResults = new EnumMap<>(Match.class);
-        for (Match match : Match.values()) {
-            matchResults.put(match, countMatch(matches, match));
-        }
-        return matchResults;
-    }
-
-    private int countMatch(Collection<Match> matches, Match type) {
-        return (int) matches.stream().filter(value -> value == type.getOpposite()).count();
-    }
-
-    public List<Gamer> getGamersCards() {
-        List<Gamer> gamers = new ArrayList<>();
-        gamers.add(dealer);
-        playerGroup.addAllTo(gamers);
-        return gamers;
+    public List<Gamer> getGamers() {
+        return gamerGroup.getGamers();
     }
 
     public int addCardToDealer() {
-        int addedCardsCount = 0;
-        while (dealer.isAddable()) {
-            dealer.addCard(cardPack.pickOne());
-            addedCardsCount++;
-        }
-
-        return addedCardsCount;
+        return gamerGroup.addCardToDealer(cardPack);
     }
 
     public void openDealerCards() {
-        dealer.openAllCards();
+        gamerGroup.openDealerCards();
     }
 }
